+++ csv/data.csv
@@ -1,0 +1,2 @@
+Name,Score
+mitko Smith,20.0
