--- new.csv
@@ -1,2 +1,0 @@
-Name,Score
-Andrey Prowless,30.0
