+++ src/main/java/com/andrewsotirov/Main.java
@@ -1,12 +1,13 @@
 package com.andrewsotirov;
 
+import com.andrewsotirov.Writer.WriteCSV;
+
 public class Main {
     public static void main(String[] args) throws Exception{
 
-        ScoreCalculator scoreCalculator = new ScoreCalculator();
-        scoreCalculator.writeAble();
+        WriteCSV writeCSV = new WriteCSV();
+        writeCSV.write();
 
     }
-
 }
